+++ Sources/Sentry/SentrySDK.m
@@ -7,6 +7,7 @@
 #import "SentryLog.h"
 #import "SentryMeta.h"
 #import "SentryScope.h"
+#import <Foundation/Foundation.h>
 
 @interface
 SentrySDK ()
@@ -28,6 +29,19 @@ static NSObject *appStartMeasurementLock;
     if (self == [SentrySDK class]) {
         appStartMeasurementLock = [[NSObject alloc] init];
     }
+}
+
+/*
+ * This method is automatically called when the library is loaded.
+ */
++ (void)load
+{
+    NSString *sentryPlistPath = [[NSBundle mainBundle] pathForResource:@"Sentry" ofType:@"plist"];
+    if (sentryPlistPath == NULL)
+        return;
+
+    NSDictionary *sentryPlist = [NSDictionary dictionaryWithContentsOfFile:sentryPlistPath];
+    [self startWithOptions:sentryPlist];
 }
 
 + (SentryHub *)currentHub
@@ -134,6 +148,11 @@ static NSObject *appStartMeasurementLock;
     SentryOptions *options = [[SentryOptions alloc] init];
     configureOptions(options);
     [SentrySDK startWithOptionsObject:options];
+}
+
++ (void)configureOptions:(void (^)(SentryOptions *options))configureOptions
+{
+    configureOptions(currentHub.client.options);
 }
 
 + (void)captureCrashEvent:(SentryEvent *)event